Prepare for possible etoc internal renamings
The macro \etocunknownlevelTF will be provided in future releases of
etoc, to check if the argument is known to etoc as a genuine sectioning
level.  This commit provides the command, so that if etoc changes its
internal naming in a future release, its version of \etocunknownlevelTF
will be the one used.  The current one works will all etoc releases at
least up to 1.1c.